f Make verify more extensible
Change verify to use ApprovalApprover rather than FileApprover. This was a very easy,
and provably safe change, as the method in question did not depend on any FileApprover
specific methods.

The concrete motivation for this change is to enable implementing
a threshold-based fuzzy image matching approver without duplicating the verify method.