B Make NamerFactory thread safe
NamerFactory uses the global variable additionalInformation to attach additional naming information to the current test run. The test initializes the variable, and uses it in a future time, so the value must remain constant for the duration of the test. While a lock may be used, probably a better solution is to give each thread its own copy of the variable, to prevent race conditions.