F!! Add auto approve reporter based on system properties
The new reporter will automatically approve the changeset (copy
received to approved) when a specified System property is set (to true
or it is simply present).

The goal is to allow to update easily all relevant files from the tests
that are run, without requiring code changes but in a controlled manner
(only when the user explicitly sets the property).